+++ test/asphalt/defsql_test.clj
@@ -26,8 +26,11 @@
   (is (= (name insert-emp) "add-new-emp")))
 
 
-(a/defsql find-employees-by-dept
-  "SELECT ^string name, ^int age, ^date joined FROM emp WHERE dept_id=^int $dept-id")
+(def sql-find-employees-by-dept "SELECT ^string name, ^int age, ^date joined FROM emp WHERE dept_id=^int $dept-id")
+(a/defsql find-employees-by-dept sql-find-employees-by-dept)
+
+
+(def sql-find-employees-by-dept-badhints "SELECT name, age, joined FROM emp WHERE dept_id=^long $dept-id")
 
 
 (a/defsql find-employees-by-level
@@ -40,3 +43,9 @@
   (is (= "SELECT name, age, joined FROM emp WHERE dept_id=? AND level IN (?, ?, ?)"
         (t/get-sql find-employees-by-level {:dept-id 20
                                             :levels [10 20 30]}))))
+
+
+(deftest test-parse-sql
+  (is (= (a/parse-sql sql-find-employees-by-dept)
+        (a/parse-sql sql-find-employees-by-dept-badhints {:param-types [:int]
+                                                          :result-types [:string :int :date]}))))